requirements.txt
pyautogui
opencv-python
python-socketio[client]
python-socketio
numpy
imutils
eventlet
pillow